SQLAlchemy for game inserts
